feat: Allow user to add a program
A user can now send a POST request to add a program.
The expected request body is in JSON format, making it easier
to share data without requiring reformatting.

---
The database pool is shared to reduce the overhead of repeatedly
opening and closing connections. However, some execution paths may
exit early, leaving connections open. For small applications like
this, it shouldn't pose an issue, but it may cause problems as the
application grows.

References:
- https://stackoverflow.com/questions/1018349/what-happens-to-an-unclosed-database-connection
- https://www.slingacademy.com/article/sqlalchemy-what-if-you-dont-close-database-connections/

---
SQL files are kept separate for easier auditing.
When the application is built, these files are embedded
into the final binary.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>