feat: Associate a client with multiple programs
Users can now associate multiple programs with a single client in a single request.
This feature modifies a separate table `client_program` from the previous ones.

A prepared statement is used instead of executing directly to reduce the
database round-trip associated with directly executing after every
request.
This also allows 0 programs to be posted instead only one.

The POST request assumes that this is the first time, so in the future
it may delete previous associations.
For now, POST and PUT/PATCH behavior are functionally the same.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>