feat: Add search functionality
A user can search client using their names.
This returns the information about the user.
For now only the name and id are returned.

SQLite FTS5 (full-text-search) is used for fast and better searches.
This will be improved over time with sanitations but for now its good
enough.
As noted, it is vulnerable to SQL injeection.
This can be mitigated by simple regex, maybe in future refactors.

References:
    https://www.sqlitetutorial.net/sqlite-full-text-search/
    https://www.bswanson.dev/blog/sqlite-full-text-search/
    https://sqlite.org/fts5.html

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>