docs: Add how to build and API information
The README now contains the documentation on how to work with the API.
Its light on the details but it can get things off the ground.
Since the project adhered to the current industry standards (whatever
that means), its relatively easy to use.

RESTful concepts and Go were used without external configurations.

Signed-off-by: Gekko Wrld <gekkowrld@gmail.com>